+++ sdks/java/io/kafka/upgrade/src/main/java/org/apache/beam/sdk/io/kafka/upgrade/KafkaIOTranslation.java
@@ -102,6 +102,7 @@ public class KafkaIOTranslation {
             .addBooleanField("allows_duplicates")
             .addNullableInt32Field("redistribute_num_keys")
             .addNullableBooleanField("offset_deduplication")
+            .addNullableBooleanField("redistribute_by_record_key")
             .addNullableLogicalTypeField("watch_topic_partition_duration", new NanosDuration())
             .addByteArrayField("timestamp_policy_factory")
             .addNullableMapField("offset_consumer_config", FieldType.STRING, FieldType.BYTES)
@@ -228,6 +229,9 @@ public class KafkaIOTranslation {
       fieldValues.put("allows_duplicates", transform.isAllowDuplicates());
       if (transform.getOffsetDeduplication() != null) {
         fieldValues.put("offset_deduplication", transform.getOffsetDeduplication());
+      }
+      if (transform.getRedistributeByRecordKey() != null) {
+        fieldValues.put("redistribute_by_record_key", transform.getRedistributeByRecordKey());
       }
       return Row.withSchema(schema).withFieldValues(fieldValues).build();
     }
@@ -361,6 +365,12 @@ public class KafkaIOTranslation {
           @Nullable Boolean offsetDeduplication = configRow.getValue("offset_deduplication");
           if (offsetDeduplication != null) {
             transform = transform.withOffsetDeduplication(offsetDeduplication);
+          }
+        }
+        if (TransformUpgrader.compareVersions(updateCompatibilityBeamVersion, "2.69.0") >= 0) {
+          @Nullable Boolean byRecordKey = configRow.getValue("redistribute_by_record_key");
+          if (byRecordKey != null) {
+            transform = transform.withRedistributeByRecordKey(byRecordKey);
           }
         }
         Duration maxReadTime = configRow.getValue("max_read_time");